Improve writer.
